Allow the synchronization method to update an existing user story

+++ Model/Epic.php
@@ -10,9 +10,9 @@ abstract class Epic implements CardInterface, EstimableInterface
     protected $title;
     /** @var string **/
     protected $status;
-    /** @var int **/
+    /** @var float **/
     protected $estimatedTime;
-    /** @var int **/
+    /** @var float **/
     protected $spentTime;
     /** @var \DateTime **/
     protected $createdAt;
@@ -60,24 +60,24 @@ abstract class Epic implements CardInterface, EstimableInterface
         return $this->status;
     }
     
-    public function setEstimatedTime(int $estimatedTime): EstimableInterface
+    public function setEstimatedTime(float $estimatedTime): EstimableInterface
     {
         $this->estimatedTime = $estimatedTime;
         
         return $this;
     }
     
-    public function getEstimatedTime(): int
+    public function getEstimatedTime(): float
     {
         return $this->getEstimatedTime();
     }
     
-    public function setSpentTime(int $spentTime): EstimableInterface
+    public function setSpentTime(float $spentTime): EstimableInterface
     {
         $this->spentTime = $spentTime;
     }
     
-    public function getSpentTime(): int
+    public function getSpentTime(): float
     {
         return $this->spentTime;
     }
