+++ Model/EstimableInterface.php
@@ -4,11 +4,11 @@ namespace Scrumban\Model;
 
 interface EstimableInterface
 {
-    public function setEstimatedTime(int $estimatedTime): self;
+    public function setEstimatedTime(float $estimatedTime): self;
     
-    public function getEstimatedTime(): int;
+    public function getEstimatedTime(): float;
     
-    public function setSpentTime(int $spentTime): self;
+    public function setSpentTime(float $spentTime): self;
     
-    public function getSpentTime(): int;
+    public function getSpentTime(): float;
 }
